+++ generate_testVectors_json.py
@@ -12,7 +12,7 @@ random.seed("mass-market-test-vectors")
 
 from protobuf_to_dict import protobuf_to_dict
 
-from massmarket_hash_event import Hasher
+from massmarket_hash_event import Hasher, store_events_pb2
 hasher = Hasher(31337, "0x0000000000000000000000001234567890abcdef")
 
 from web3 import Account
@@ -29,7 +29,8 @@ def debug(message):
 def hex(b):
   return binascii.hexlify(b).decode('utf-8')
 
-import schema_pb2
+def unhex(a):
+    return binascii.a2b_hex(a[2:])
 
 kc1 = Account.from_key(random.randbytes(32))
 debug(f"kc1: {kc1.address}")
@@ -38,39 +39,39 @@ debug(f"kc2: {kc2.address}")
 
 events = []
 
-manifest = schema_pb2.StoreManifest()
+manifest = store_events_pb2.StoreManifest()
 manifest.event_id = random.randbytes(32)
 manifest.store_token_id = random.randbytes(32)
 manifest.domain = "https://masslbs.xyz"
 manifest.published_tag_id = random.randbytes(32)
 events.append(manifest)
 
-newKc1 = schema_pb2.NewKeyCard()
+newKc1 = store_events_pb2.NewKeyCard()
 newKc1.event_id = random.randbytes(32)
 newKc1.card_public_key = public_key_from_account(kc1).to_bytes()
 newKc1.user_wallet_addr = random.randbytes(20)
 events.append(newKc1)
 
-newKc2 = schema_pb2.NewKeyCard()
+newKc2 = store_events_pb2.NewKeyCard()
 newKc2.event_id = random.randbytes(32)
 newKc2.card_public_key = public_key_from_account(kc2).to_bytes()
 newKc2.user_wallet_addr = random.randbytes(20)
 events.append(newKc2)
 
-update2 = schema_pb2.UpdateManifest()
+update2 = store_events_pb2.UpdateStoreManifest()
 update2.event_id = random.randbytes(32)
-update2.field = schema_pb2.UpdateManifest.ManifestField.MANIFEST_FIELD_DOMAIN
+update2.field = store_events_pb2.UpdateStoreManifest.ManifestField.MANIFEST_FIELD_DOMAIN
 update2.string = "https://store.mass.market"
 events.append(update2)
 
-newTag1 = schema_pb2.CreateTag()
+newTag1 = store_events_pb2.CreateTag()
 newTag1.event_id = random.randbytes(32)
 newTag1.name = "Tag1"
 events.append(newTag1)
 
-update3 = schema_pb2.UpdateManifest()
+update3 = store_events_pb2.UpdateStoreManifest()
 update3.event_id = random.randbytes(32)
-update3.field = schema_pb2.UpdateManifest.ManifestField.MANIFEST_FIELD_PUBLISHED_TAG
+update3.field = store_events_pb2.UpdateStoreManifest.ManifestField.MANIFEST_FIELD_PUBLISHED_TAG
 update3.tag_id = newTag1.event_id
 events.append(update3)
 
@@ -78,192 +79,214 @@ events.append(update3)
 erc20_one = random.randbytes(20)
 erc20_two = random.randbytes(20)
 
-addErc20One = schema_pb2.UpdateManifest()
+addErc20One = store_events_pb2.UpdateStoreManifest()
 addErc20One.event_id = random.randbytes(32)
-addErc20One.field = schema_pb2.UpdateManifest.ManifestField.MANIFEST_FIELD_ADD_ERC20
+addErc20One.field = store_events_pb2.UpdateStoreManifest.ManifestField.MANIFEST_FIELD_ADD_ERC20
 addErc20One.erc20_addr = erc20_one
 events.append(addErc20One)
 
-addErc20Two = schema_pb2.UpdateManifest()
+addErc20Two = store_events_pb2.UpdateStoreManifest()
 addErc20Two.event_id = random.randbytes(32)
-addErc20Two.field = schema_pb2.UpdateManifest.ManifestField.MANIFEST_FIELD_ADD_ERC20
+addErc20Two.field = store_events_pb2.UpdateStoreManifest.ManifestField.MANIFEST_FIELD_ADD_ERC20
 addErc20Two.erc20_addr = erc20_two
 events.append(addErc20Two)
 
-rmTokenOne = schema_pb2.UpdateManifest()
+rmTokenOne = store_events_pb2.UpdateStoreManifest()
 rmTokenOne.event_id = random.randbytes(32)
-rmTokenOne.field = schema_pb2.UpdateManifest.ManifestField.MANIFEST_FIELD_REMOVE_ERC20
+rmTokenOne.field = store_events_pb2.UpdateStoreManifest.ManifestField.MANIFEST_FIELD_REMOVE_ERC20
 rmTokenOne.erc20_addr = erc20_one
 events.append(rmTokenOne)
 
 # listing managment
-newItem = schema_pb2.CreateItem()
+newItem = store_events_pb2.CreateItem()
 newItem.event_id = random.randbytes(32)
 newItem.price = "42.00"
 newItem.metadata = b'{"title":"best schoes", "image":"https://masslbs.xyz/schoes.jpg"}'
 events.append(newItem)
 
-publishItem = schema_pb2.AddToTag()
+publishItem = store_events_pb2.UpdateTag()
 publishItem.event_id = random.randbytes(32)
 publishItem.tag_id = newTag1.event_id
+publishItem.action = store_events_pb2.UpdateTag.TAG_ACTION_ADD_ITEM
 publishItem.item_id = newItem.event_id
 events.append(publishItem)
 
-changePrice = schema_pb2.UpdateItem()
+changePrice = store_events_pb2.UpdateItem()
 changePrice.event_id = random.randbytes(32)
 changePrice.item_id = newItem.event_id
-changePrice.field = schema_pb2.UpdateItem.ItemField.ITEM_FIELD_PRICE
+changePrice.field = store_events_pb2.UpdateItem.ItemField.ITEM_FIELD_PRICE
 changePrice.price = "23.00"
 events.append(changePrice)
 
-notPublishedItem = schema_pb2.CreateItem()
+notPublishedItem = store_events_pb2.CreateItem()
 notPublishedItem.event_id = random.randbytes(32)
 notPublishedItem.price = "13.12"
 notPublishedItem.metadata = b'{"title":"not yet published", "image":"https://masslbs.xyz/schoes.jpg"}'
 events.append(notPublishedItem)
 
-publishItem2 = schema_pb2.AddToTag()
+publishItem2 = store_events_pb2.UpdateTag()
 publishItem2.event_id = random.randbytes(32)
 publishItem2.tag_id = newTag1.event_id
+publishItem2.action = store_events_pb2.UpdateTag.TAG_ACTION_ADD_ITEM
 publishItem2.item_id = notPublishedItem.event_id
 events.append(publishItem2)
 
-unpublishItem = schema_pb2.RemoveFromTag()
+unpublishItem = store_events_pb2.UpdateTag()
 unpublishItem.event_id = random.randbytes(32)
 unpublishItem.tag_id = newTag1.event_id
+unpublishItem.action = store_events_pb2.UpdateTag.TAG_ACTION_REMOVE_ITEM
 unpublishItem.item_id = notPublishedItem.event_id
 events.append(unpublishItem)
 
-changeStock = schema_pb2.ChangeStock(item_ids = [newItem.event_id, notPublishedItem.event_id], diffs = [100, 123] )
+changeStock = store_events_pb2.ChangeStock(item_ids = [newItem.event_id, notPublishedItem.event_id], diffs = [100, 123] )
 changeStock.event_id = random.randbytes(32)
 events.append(changeStock)
 
-# open cart
-cart1 = schema_pb2.CreateCart()
-cart1.event_id = random.randbytes(32)
-events.append(cart1)
+# open order
+order1 = store_events_pb2.CreateOrder()
+order1.event_id = random.randbytes(32)
+events.append(order1)
 
-atc1 = schema_pb2.ChangeCart()
+change = store_events_pb2.UpdateOrder.ChangeItems()
+change.item_id = newItem.event_id
+change.quantity = 23
+atc1 = store_events_pb2.UpdateOrder(change_items=change)
 atc1.event_id = random.randbytes(32)
-atc1.cart_id = cart1.event_id
-atc1.item_id = newItem.event_id
-atc1.quantity = 23
+atc1.order_id = order1.event_id
 events.append(atc1)
 
-# cart to be payed
-cart2 = schema_pb2.CreateCart()
-cart2.event_id = random.randbytes(32)
-events.append(cart2)
+# order to be payed
+order2 = store_events_pb2.CreateOrder()
+order2.event_id = random.randbytes(32)
+events.append(order2)
 
-atc2 = schema_pb2.ChangeCart()
+change2 = store_events_pb2.UpdateOrder.ChangeItems()
+change2.item_id = newItem.event_id
+change2.quantity = 42
+atc2 = store_events_pb2.UpdateOrder(change_items=change2)
 atc2.event_id = random.randbytes(32)
-atc2.cart_id = cart2.event_id
-atc2.item_id = newItem.event_id
-atc2.quantity = 42
+atc2.order_id = order2.event_id
 events.append(atc2)
 
 # created by the relay.
-# client would send a CommitCartRequest to finalize the cart.
-commit_cart = schema_pb2.CartFinalized()
-commit_cart.event_id = random.randbytes(32)
-commit_cart.cart_id = cart2.event_id
-commit_cart.purchase_addr = random.randbytes(20)
-commit_cart.erc20_addr = random.randbytes(20)
-commit_cart.sub_total = "1764.00" # 42*42
-commit_cart.sales_tax = "88.20"   # 5%
-commit_cart.total = "1852.20"
-commit_cart.total_in_crypto = "1852.20"
-events.append(commit_cart)
+# client would send a CommitOrderRequest to finalize the order.
+commit_order = store_events_pb2.UpdateOrder.ItemsFinalized()
+commit_order.ttl = "23"
+commit_order.payee_addr = unhex(kc1.address)
+commit_order.payment_id = random.randbytes(32)
+commit_order.currency_addr = random.randbytes(20)
+commit_order.order_hash = random.randbytes(32)
+commit_order.is_payment_endpoint = True
+commit_order.shop_signature = random.randbytes(64)
+commit_order.sub_total = "1764.00" # 42*42
+commit_order.sales_tax = "88.20"   # 5%
+commit_order.total = "1852.20"
+commit_order.total_in_crypto = "1852.20"
+
+update_order = store_events_pb2.UpdateOrder(items_finalized=commit_order)
+update_order.event_id = random.randbytes(32)
+update_order.order_id = order2.event_id
+events.append(update_order)
 
 # would be created by the relay after payment is completed
-payedCart = schema_pb2.ChangeStock(item_ids=[newItem.event_id], diffs=[-42])
-payedCart.event_id = random.randbytes(32)
-payedCart.cart_id = cart2.event_id
-payedCart.tx_hash = random.randbytes(32)
-events.append(payedCart)
+payedOrder = store_events_pb2.ChangeStock(item_ids=[newItem.event_id], diffs=[-42])
+payedOrder.event_id = random.randbytes(32)
+payedOrder.order_id = order2.event_id
+payedOrder.tx_hash = random.randbytes(32)
+events.append(payedOrder)
 
-# this cart will be abandoned
-cart3 = schema_pb2.CreateCart()
-cart3.event_id = random.randbytes(32)
-events.append(cart3)
+# this order will be abandoned
+order3 = store_events_pb2.CreateOrder()
+order3.event_id = random.randbytes(32)
+events.append(order3)
 
-atc3 = schema_pb2.ChangeCart()
+change3 = store_events_pb2.UpdateOrder.ChangeItems()
+change3.item_id = newItem.event_id
+change3.quantity = 1
+atc3 = store_events_pb2.UpdateOrder(change_items=change3)
 atc3.event_id = random.randbytes(32)
-atc3.cart_id = cart3.event_id
-atc3.item_id = newItem.event_id
-atc3.quantity = 1
+atc3.order_id = order3.event_id
 events.append(atc3)
 
-commit_cart3 = schema_pb2.CartFinalized()
-commit_cart3.event_id = random.randbytes(32)
-commit_cart3.cart_id = cart3.event_id
-commit_cart3.purchase_addr = random.randbytes(20)
-commit_cart3.sub_total = "42.00"
-commit_cart3.sales_tax = "2.10"
-commit_cart3.total = "44.10"
-commit_cart3.total_in_crypto = "xxxx"
-events.append(commit_cart3)
+commit_order3 = store_events_pb2.UpdateOrder.ItemsFinalized()
+commit_order3.ttl = "1"
+commit_order3.payee_addr = unhex(kc1.address)
+commit_order3.payment_id = random.randbytes(32)
+commit_order3.order_hash = random.randbytes(32)
+commit_order3.is_payment_endpoint = False
+commit_order3.shop_signature = random.randbytes(64)
+commit_order3.sub_total = "42.00"
+commit_order3.sales_tax = "2.10"
+commit_order3.total = "44.10"
+commit_order3.total_in_crypto = "xxxx"
+update_order2 = store_events_pb2.UpdateOrder(items_finalized=commit_order3)
+update_order2.event_id = random.randbytes(32)
+update_order2.order_id = order3.event_id
+events.append(update_order2)
 
 # 24hrs pass and the sale times out
-cart_abandoned = schema_pb2.CartAbandoned()
-cart_abandoned.event_id = random.randbytes(32)
-cart_abandoned.cart_id = cart3.event_id
-events.append(cart_abandoned)
+cancel = store_events_pb2.UpdateOrder.OrderCanceled(timestamp=23)
+update_order3 = store_events_pb2.UpdateOrder(order_canceled=cancel)
+update_order3.event_id = random.randbytes(32)
+update_order3.order_id = order3.event_id
+events.append(update_order3)
 
-# cart 4 is in limbo. Finalized but not yet payed
-cart4 = schema_pb2.CreateCart()
-cart4.event_id = random.randbytes(32)
-events.append(cart4)
+# order 4 is in limbo. Finalized but not yet payed
+order4 = store_events_pb2.CreateOrder()
+order4.event_id = random.randbytes(32)
+events.append(order4)
 
-atc3 = schema_pb2.ChangeCart()
-atc3.event_id = random.randbytes(32)
-atc3.cart_id = cart4.event_id
-atc3.item_id = newItem.event_id
-atc3.quantity = 4
-events.append(atc3)
+change4 = store_events_pb2.UpdateOrder.ChangeItems()
+change4.item_id = newItem.event_id
+change4.quantity = 4
+atc4 = store_events_pb2.UpdateOrder(change_items=change4)
+atc4.event_id = random.randbytes(32)
+atc4.order_id = order4.event_id
+events.append(atc4)
 
-commit_cart4 = schema_pb2.CartFinalized()
-commit_cart4.event_id = random.randbytes(32)
-commit_cart4.cart_id = cart4.event_id
-commit_cart4.purchase_addr = random.randbytes(20)
-commit_cart4.sub_total = "168.00"
-commit_cart4.sales_tax = "8.40"
-commit_cart4.total = "176.40"
-commit_cart4.total_in_crypto = "xxxx"
-events.append(commit_cart4)
+commit_order4 = store_events_pb2.UpdateOrder.ItemsFinalized()
+commit_order4.ttl = "1"
+commit_order4.payee_addr = unhex(kc1.address)
+commit_order4.is_payment_endpoint = False
+commit_order4.payment_id = random.randbytes(32)
+commit_order4.order_hash = random.randbytes(32)
+commit_order4.shop_signature = random.randbytes(64)
+commit_order4.sub_total = "168.00"
+commit_order4.sales_tax = "8.40"
+commit_order4.total = "176.40"
+commit_order4.total_in_crypto = "xxxx"
+update_order3 = store_events_pb2.UpdateOrder(items_finalized=commit_order4)
+update_order3.event_id = random.randbytes(32)
+update_order3.order_id = order4.event_id
+events.append(update_order3)
 
 wrapped_events = []
 for idx, evt in enumerate(events):
-  debug("\nEvent: {}".format(idx))
+  type_name = evt.__class__.__name__
+
+  debug(f"\nEvent idx={idx} type={type_name}\n{evt}")
 
   wrapped = None
-  type_name = evt.__class__.__name__
   if type_name == "StoreManifest":
-    wrapped = schema_pb2.Event(store_manifest=evt)
-  elif type_name == "UpdateManifest":
-    wrapped = schema_pb2.Event(update_manifest=evt)
+    wrapped = store_events_pb2.StoreEvent(store_manifest=evt)
+  elif type_name == "UpdateStoreManifest":
+    wrapped = store_events_pb2.StoreEvent(update_store_manifest=evt)
   elif type_name == "CreateItem":
-    wrapped = schema_pb2.Event(create_item=evt)
-  elif type_name == "CreateTag":
-    wrapped = schema_pb2.Event(create_tag=evt)
+    wrapped = store_events_pb2.StoreEvent(create_item=evt)
   elif type_name == "UpdateItem":
-    wrapped = schema_pb2.Event(update_item=evt)
-  elif type_name == "AddToTag":
-    wrapped = schema_pb2.Event(add_to_tag=evt)
-  elif type_name == "RemoveFromTag":
-    wrapped = schema_pb2.Event(remove_from_tag=evt)
+    wrapped = store_events_pb2.StoreEvent(update_item=evt)
+  elif type_name == "CreateTag":
+    wrapped = store_events_pb2.StoreEvent(create_tag=evt)
+  elif type_name == "UpdateTag":
+    wrapped = store_events_pb2.StoreEvent(update_tag=evt)
   elif type_name == "NewKeyCard":
-    wrapped = schema_pb2.Event(new_key_card=evt)
+    wrapped = store_events_pb2.StoreEvent(new_key_card=evt)
   elif type_name == "ChangeStock":
-    wrapped = schema_pb2.Event(change_stock=evt)
-  elif type_name == "CreateCart":
-    wrapped = schema_pb2.Event(create_cart=evt)
-  elif type_name == "ChangeCart":
-    wrapped = schema_pb2.Event(change_cart=evt)
-  elif type_name == "CartFinalized":
-    wrapped = schema_pb2.Event(cart_finalized=evt)
-  elif type_name == "CartAbandoned":
-    wrapped = schema_pb2.Event(cart_abandoned=evt)
+    wrapped = store_events_pb2.StoreEvent(change_stock=evt)
+  elif type_name == "CreateOrder":
+    wrapped = store_events_pb2.StoreEvent(create_order=evt)
+  elif type_name == "UpdateOrder":
+    wrapped = store_events_pb2.StoreEvent(update_order=evt)
   else:
     raise Exception(f"Unknown event type: {type_name}")
 
@@ -326,15 +349,15 @@ output = {
     # an array of items published
     "published_items": [hex(publishItem.item_id)],
 
-    "open_carts": {
-      hex(cart1.event_id): {
+    "open_orders": {
+      hex(order1.event_id): {
           hex(newItem.event_id): 23
       }
     },
 
-    "commited_carts": {
-      hex(cart4.event_id): {
-          "purchase_addr": hex(commit_cart4.purchase_addr),
+    "commited_orders": {
+      hex(order4.event_id): {
+          "payment_id": hex(commit_order4.payment_id),
           "items": {
               hex(newItem.event_id): 4
           },
@@ -342,13 +365,13 @@ output = {
       },
     },
 
-    "payed_carts": [{
-        "cart_id": hex(cart2.event_id),
-        "tx_hash": hex(payedCart.tx_hash),
+    "payed_orders": [{
+        "order_id": hex(order2.event_id),
+        "tx_hash": hex(payedOrder.tx_hash),
     }],
 
-    "abandoned_carts": [
-        hex(cart3.event_id),
+    "abandoned_orders": [
+        hex(order3.event_id),
     ],
 
     "inventory": {